Add command line option to set display orientation
The problem: currently the display orientation (aka rotation) is set in
the profile. XCSoar will rotate the screen according to
"DisplayOrientation" setting soon as profile is read. However, there are
several UIs that are shown before the profile is loaded. Notably, the
initial "Fly/Simulator" screen and the actual profile selector. XCSoar
would always render these screens in default "landscape" orientation.
This causes problems for some devices with fixed installation (such as
Openvario), on which these initial screens might appear "sideways".

The solution: add a command line option, "-orientation=N" on devices
that support the software screen rotation. The option would allow to
provide XCSoar with the initial screen orientation.